refactor: move load_static back to xblock
Fixes an issue with pathing of resource_string
